Don't fail when presented with blob types
The server does not allow submitting data for a table with a non-null blob
column since it's (right now) unnecessary and introduces a potential security
risk. Only 'null' is able to be submitted for blob columns (if and only if that
column is nullable). The dynamic form generator takes this into account and
automatically disables non-null blob inputs.